+++ Pandora/Paradigm/Primary/Linear.hs
@@ -1,3 +1,4 @@
 module Pandora.Paradigm.Primary.Linear (module Exports) where
 
+import Pandora.Paradigm.Primary.Linear.Matrix as Exports
 import Pandora.Paradigm.Primary.Linear.Vector as Exports