+++ Pandora/Paradigm/Primary/Linear/Matrix.hs
@@ -1,0 +1,5 @@
+module Pandora.Paradigm.Primary.Linear.Matrix where
+
+import Pandora.Paradigm.Primary.Linear.Vector (Vector)
+
+newtype Matrix i j a = Matrix (Vector i (Vector j a))